Cleaning up the various input controllers interfaces
