--- Frontend/library/src/Inputs/ITouchController.ts
@@ -1,8 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-/**
- * Interface for Touch Events
- */
-export interface ITouchController {
-    unregisterTouchEvents(): void;
-}
